Split the mixed pilot test files into the two lanes
Three pilot test files hold both window-driving and headless classes, so
the move that grouped the window tests under tests/gui had to leave them
behind. Each one still costs a pull request the slow half of its run.

Split them along the same line the directory draws. The classes guarded
by NO_LIVE_WINDOW go to tests/gui/test_painter_layers.py,
tests/gui/test_shortcut.py, and tests/gui/test_solution_info.py; the
classes that build widgets without mapping a window stay where they are.
The skip guards, the test bodies, and their comments are untouched.

A split file cannot simply move its imports, so each new file carries its
own copy of the machinery its classes need. The _click helper, the
_PANEL_CHORDS and _DRAW_TOOL_KEYS tables are duplicated because both
sides call them; a shared import between tests/ and tests/gui/ would
couple the two lanes the directory exists to keep apart. What no
remaining class uses, including the now dangling NO_LIVE_WINDOW
definitions, is deleted from the source files.

Related to #1222
